+++ src/connect.lua
@@ -1,7 +1,25 @@
+--!strict
 local Roact = require(script.Parent.Parent.Roact)
+
 local shallowEqual = require(script.Parent.shallowEqual)
 local join = require(script.Parent.join)
 local StoreContext = require(script.Parent.StoreContext)
+
+local types = require(script.Parent.types)
+
+type ThunkfulDispatchProp<State> = types.ThunkfulDispatchProp<State>
+type MapStateToProps<StoreState, Props, PartialProps> = types.MapStateToProps<StoreState, Props, PartialProps>
+type MapStateToPropsOrThunk<StoreState, Props, PartialProps> = types.MapStateToPropsOrThunk<
+	StoreState,
+	Props,
+	PartialProps
+>
+type ActionCreatorMap = types.ActionCreatorMap
+type MapDispatchToProps<StoreState, PartialProps> = types.MapDispatchToProps<StoreState, PartialProps>
+type MapDispatchToPropsOrActionCreator<StoreState, PartialProps> = types.MapDispatchToPropsOrActionCreator<
+	StoreState,
+	PartialProps
+>
 
 --[[
 	Formats a multi-line message with printf-style placeholders.
@@ -12,6 +30,15 @@ end
 
 local function noop()
 	return nil
+end
+
+--[[
+	Returns `true` if the value can be called i.e. you can write `value(...)`.
+]]
+local function isCallable(value: any): boolean
+	return type(value) == "function"
+		or (type(value) == "table" and getmetatable(value) and getmetatable(value).__call ~= nil)
+		or false
 end
 
 --[[
@@ -47,11 +74,19 @@ end
 		() -> (storeState, props) -> partialProps
 	mapDispatchToProps: (dispatch) -> partialProps
 ]]
-local function connect(mapStateToPropsOrThunk, mapDispatchToProps)
-	local connectTrace = debug.traceback()
-
+local function connect<StoreState, Props, MappedStatePartialProps, MappedDispatchPartialProps>(
+	mapStateToPropsOrThunk: MapStateToPropsOrThunk<
+		StoreState,
+		Props,
+		MappedStatePartialProps
+	>?,
+	mapDispatchToProps: MapDispatchToPropsOrActionCreator<
+		StoreState,
+		MappedDispatchPartialProps
+	>?
+)
 	if mapStateToPropsOrThunk ~= nil then
-		assert(typeof(mapStateToPropsOrThunk) == "function", "mapStateToProps must be a function or nil!")
+		assert(isCallable(mapStateToPropsOrThunk), "mapStateToProps must be a function or nil!")
 	else
 		mapStateToPropsOrThunk = noop
 	end
@@ -68,6 +103,7 @@ local function connect(mapStateToPropsOrThunk, mapDispatchToProps)
 
 	return function(innerComponent)
 		if innerComponent == nil then
+			local connectTrace = debug.traceback()
 			local message = formatMessage({
 				"connect returns a function that must be passed a component.",
 				"Check the connection at:",
@@ -87,23 +123,7 @@ local function connect(mapStateToPropsOrThunk, mapDispatchToProps)
 			if prevState.stateUpdater ~= nil then
 				return prevState.stateUpdater(nextProps.innerProps, prevState)
 			end
-		end
-
-		function Connection:createStoreConnection()
-			self.storeChangedConnection = self.store.changed:connect(function(storeState)
-				self:setState(function(prevState, props)
-					local mappedStoreState = prevState.mapStateToProps(storeState, props.innerProps)
-
-					-- We run this check here so that we only check shallow
-					-- equality with the result of mapStateToProps, and not the
-					-- other props that could be passed through the connector.
-					if shallowEqual(mappedStoreState, prevState.mappedStoreState) then
-						return nil
-					end
-
-					return prevState.stateUpdater(props, prevState, mappedStoreState)
-				end)
-			end)
+			return nil
 		end
 
 		function Connection:init(props)
@@ -123,15 +143,16 @@ local function connect(mapStateToPropsOrThunk, mapDispatchToProps)
 
 			local storeState = self.store:getState()
 
-			local mapStateToProps = mapStateToPropsOrThunk
+			local mapStateToProps =
+				mapStateToPropsOrThunk :: MapStateToProps<StoreState, Props, MappedStatePartialProps>
 			local mappedStoreState = mapStateToProps(storeState, self.props.innerProps)
 
 			-- mapStateToPropsOrThunk can return a function instead of a state
 			-- value. In this variant, we keep that value as mapStateToProps
 			-- instead of the original mapStateToProps. This matches react-redux
 			-- and enables connectors to keep instance-level state.
-			if typeof(mappedStoreState) == "function" then
-				mapStateToProps = mappedStoreState
+			if isCallable(mappedStoreState) then
+				mapStateToProps = mappedStoreState :: any
 				mappedStoreState = mapStateToProps(storeState, self.props.innerProps)
 			end
 
@@ -151,19 +172,21 @@ local function connect(mapStateToPropsOrThunk, mapDispatchToProps)
 				return self.store:dispatch(...)
 			end
 
-			local mappedStoreDispatch
-			if mapDispatchType == "table" then
+			local mappedStoreDispatch: any
+			if isCallable(mapDispatchToProps) then
+				mappedStoreDispatch = (mapDispatchToProps :: MapDispatchToProps<StoreState, MappedDispatchPartialProps>)(
+					(dispatch :: any) :: ThunkfulDispatchProp<StoreState>
+				)
+			elseif mapDispatchType == "table" then
 				mappedStoreDispatch = {}
 
-				for key, actionCreator in pairs(mapDispatchToProps) do
-					assert(typeof(actionCreator) == "function", "mapDispatchToProps must contain function values")
+				for key, actionCreator in pairs(mapDispatchToProps :: ActionCreatorMap) do
+					assert(isCallable(actionCreator), "mapDispatchToProps must contain function values")
 
 					mappedStoreDispatch[key] = function(...)
 						dispatch(actionCreator(...))
 					end
 				end
-			elseif mapDispatchType == "function" then
-				mappedStoreDispatch = mapDispatchToProps(dispatch)
 			end
 
 			local stateUpdater = makeStateUpdater(self.store)
@@ -191,12 +214,37 @@ local function connect(mapStateToPropsOrThunk, mapDispatchToProps)
 			for key, value in pairs(extraState) do
 				self.state[key] = value
 			end
+		end
 
-			self:createStoreConnection()
+		function Connection:didMount()
+			local updateStateWithStore = function(storeState)
+				self:setState(function(prevState, props)
+					local mappedStoreState = prevState.mapStateToProps(storeState, props.innerProps)
+
+					-- We run this check here so that we only check shallow
+					-- equality with the result of mapStateToProps, and not the
+					-- other props that could be passed through the connector.
+					if shallowEqual(mappedStoreState, prevState.mappedStoreState) then
+						return nil
+					end
+
+					return prevState.stateUpdater(props.innerProps, prevState, mappedStoreState)
+				end)
+			end
+
+			-- Update store state on mount to catch missed state updates between
+			-- init and mount. State could be stale otherwise.
+			updateStateWithStore(self.store:getState())
+
+			-- Connect state updater to the Rodux store
+			self.storeChangedConnection = self.store.changed:connect(updateStateWithStore)
 		end
 
 		function Connection:willUnmount()
-			self.storeChangedConnection:disconnect()
+			if self.storeChangedConnection then
+				self.storeChangedConnection:disconnect()
+				self.storeChangedConnection = nil
+			end
 		end
 
 		function Connection:render()
@@ -212,7 +260,7 @@ local function connect(mapStateToPropsOrThunk, mapDispatchToProps)
 						innerProps = self.props,
 						store = store,
 					})
-				end
+				end,
 			})
 		end